requirements.txt
Django==5.0.3
google-generativeai==0.4.1
mtranslate==1.8
pillow==10.2.0
unidecode==1.3.8
djangorestframework==3.15.1
django-filter==24.2
drf-yasg==1.21.7
python-magic-bin==0.4.14
python-dotenv==1.0.1
